add database service && fix bugs
