+++ server/src/auth/entities/session.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, JoinColumn, OneToOne } from 'typeorm'
 import { StaffEntity } from '@/modules/staff/entities'
-import { BaseEntity } from '@core/database'
+import { BaseEntity } from '@/core/database/entity'
 
 @Entity('Session')
 export class SessionEntity extends BaseEntity {
