+++ server/src/core/database/database.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common'
+
+import { DataBaseService } from './database.service'
+
+@Module({
+	imports: [],
+	exports: [DataBaseService],
+	providers: [DataBaseService]
+})
+export class DataBaseModule {}
