+++ server/src/core/database/entity/base.entity.ts
