+++ server/src/modules/club/club.module.ts
@@ -4,9 +4,10 @@ import { ClubService } from './club.service'
 import { TypeOrmModule } from '@nestjs/typeorm'
 import { ClubEntity } from './entities'
 import { StaffModule } from '../staff/staff.module'
+import { DataBaseModule } from '@/core/database/database.module'
 
 @Module({
-	imports: [TypeOrmModule.forFeature([ClubEntity]), StaffModule],
+	imports: [TypeOrmModule.forFeature([ClubEntity]), StaffModule, DataBaseModule],
 	controllers: [ClubController],
 	providers: [ClubService],
 	exports: [ClubService]
